+++ packages/edit-site/src/components/sidebar/settings-header/style.scss
